+++ core-ng/src/main/java/core/framework/module/APIClientConfig.java
@@ -1,5 +1,6 @@
 package core.framework.module;
 
+import core.framework.internal.inject.InjectValidator;
 import core.framework.web.service.WebServiceClientInterceptor;
 import core.framework.web.service.WebServiceClientProxy;
 
@@ -14,6 +15,7 @@ public final class APIClientConfig {     // returned from APIConfig, must be pub
     }
 
     public void intercept(WebServiceClientInterceptor interceptor) {
+        new InjectValidator(interceptor).validate();
         client.intercept(interceptor);
     }
 }